+++ clodoo/clodoo/scripts/__init__.py
@@ -1,2 +1,1 @@
 from . import main
-from . import list_requirements