+++ wok_code/wok_code/scripts/dist_pkg.py
@@ -7,23 +7,16 @@ import sys
 def main(cli_args=None):
     if not cli_args:
         cli_args = sys.argv[1:]
-    cmd = os.path.abspath(
-        os.path.join(
-            os.path.dirname(__file__), '%s.sh' % os.path.basename(__file__)[0:-3]
-        )
-    )
+    cmd = '%s.sh' % os.path.splitext(os.path.abspath(__file__))[0]
     if not os.path.isfile(cmd):
-        cmd = os.path.abspath(
-            os.path.join(
-                os.path.dirname(__file__),
-                '..',
-                '%s.sh' % os.path.basename(__file__)[0:-3]
-            )
-        )
+        cmd = os.path.split(cmd)
+        cmd = os.path.join(os.path.dirname(cmd[0]), cmd[1])
     if not os.path.isfile(cmd):
         print('Internal package error: file %s not found!' % cmd)
     for arg in cli_args:
-        if ' ' in arg:
+        if '<' in arg or '>' in arg:
+            arg = "'%s'" % arg.replace("'", r"\'")
+        elif ' ' in arg:
             if '"' in arg:
                 arg = '"%s"' % arg.replace('"', r'\"')
             else:
@@ -36,3 +29,7 @@ def main(cli_args=None):
             arg = '%s' % arg
         cmd = '%s %s' % (cmd, arg)
     return os.system(cmd)
+
+
+if __name__ == "__main__":
+    exit(main())