+++ odoo_score/odoo_score/egg-info/history.rst
@@ -1,3 +1,9 @@
+2.0.0.2 (2022-09-22)
+~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] run_odo_debug: test with random rpcport for multiple tests
+* [IMP] odoo_shell.py: new actions
+
 2.0.0.1 (2022-09-07)
 ~~~~~~~~~~~~~~~~~~~~