+++ python_plus/python_plus/egg-info/history.rst
@@ -1,7 +1,7 @@
 2.0.0.3 (2022-09-14)
 ~~~~~~~~~~~~~~~~~~~~
 
-* [FIX] vem: install package with > and < filter
+* [FIX] vem: install package with list_requirements.py
 
 2.0.0.2 (2022-09-10)
 ~~~~~~~~~~~~~~~~~~~~